django_empty/requirements.txt
django
djangorestframework==3.3.3
